Move WP version of Framework\Configurations functions into the base class
